refactor(channels): 将 ChannelModalContent 和 SecretModalContent 移至 channels 目录
